SUCCESS
1. Migrated all code from BirdListActivityMain to BirdListFragment (the whole purpose of using fragments)

2. BirdListActivityMain now extends from SingleFragmentActivity

3. Fixed some small issues with creating the wrong view layout
